ruby version (see description)
Make gemfile read from .ruby-version which should stay in sync. you can
also specify ruby "3.4.3" in your gemfile, but build tools and local
dev/editors will look for .ruby-version. this is a way to keep both in
sync so you dont hav to change them separately. cloud66 and heroku
expect the ruby version in the gemfile