simplify core fields
this will prevent unecessary confusion on the frontend when having to
specify "primary_phone" in like 1000 places. imo it should be core
simple fields (phone, address) but if there is more nuance, then it can
be mobile_phone, work_phone etc, because those fields very likely will
never be in the frontend, instead, they will be MAPPED to those fields
FROM your phone field. a buyer may want 13 different phone fields ,but
pete is never going to want to collect multiple phones, so we just take
"phone" and pass it to those 13 fields